blockplain_rag/requirements.txt
pathway
fastapi
uvicorn
sentence-transformers
python-dotenv
websockets
aiohttp
jinja2
numpy
pydantic